Resolve(6/6): Discussion — relocate use-comment-url out of documents
Discussion lives in `comments` (shared) but explorations made it import from
`documents` (feature): setHoveredChildTargetId and useCommentUrl. That breaks
the boundary, and `comments` has no host seam.

Resolution needs no seam, because neither dependency is truly document-specific:
- setHoveredChildTargetId was imported but never used -> drop it.
- useCommentUrl only depends on metabase/router + a comments type; it was simply
  mis-filed. Move it from documents/hooks to comments/hooks. All importers
  (Discussion, CommentsMenu, documentEditorHost) now pull it from comments
  (shared), which every tier may import. useDispatch became unused -> drop it.

Takeaway for the team: not everything currently under documents/ is
document-specific. Hooks with no document dependency should live in a shared
module so comments/editor can use them without crossing the boundary.
