+++ frontend/src/metabase/comments/components/Discussion/Discussion.tsx
@@ -9,9 +9,8 @@ import {
 } from "metabase/api";
 import { getCommentNodeId } from "metabase/comments/utils";
 import { useToast } from "metabase/common/hooks";
-import { setHoveredChildTargetId } from "metabase/documents/documents.slice";
-import { useCommentUrl } from "metabase/documents/hooks/use-comment-url";
-import { useDispatch, useSelector } from "metabase/redux";
+import { useCommentUrl } from "metabase/comments/hooks/use-comment-url";
+import { useSelector } from "metabase/redux";
 import { getUser } from "metabase/selectors/user";
 import { Avatar, Stack, Timeline, rem } from "metabase/ui";
 import type {
